refactor: modify discord channel model

+++ models/discord_channel.go
@@ -11,7 +11,7 @@ type DiscordChannel struct {
 	Name        string    `db:"name"`
 	ChannelId   string    `db:"channel_id,unique"`
 	ServerId    string    `db:"server_id,unique"`
-	StreamLimit int64     `db:"stream_limit"`
+	StreamLimit int       `db:"stream_limit"`
 	UserID      uuid.UUID `db:"user_id,fk"`
 	CreatedAt   time.Time `db:"created_at"`
 	UpdatedAt   time.Time `db:"updated_at"`
